refactor(gpu-video): fold VAAPI DMA-BUF backend into gpu-video
