+++ Cargo.toml
@@ -15,7 +15,6 @@ members = [
     # Libraries
     "decklink",
     "gpu-video",
-    "va-video",
     "libcef",
     "rtmp",
 
@@ -40,7 +39,6 @@ smelter-core = { path = "smelter-core" }
 smelter-render = { path = "smelter-render" }
 libcef = { path = "libcef" }
 gpu-video = { path = "gpu-video", features = ["expose-parsers"] }
-va-video = { path = "va-video" }
 rtmp = { path = "rtmp" }
 bytes = "1.11.1"
 serde_json = { version = "1.0.99", features = ["preserve_order"] }
