+++ gpu-video/Cargo.toml
@@ -21,6 +21,8 @@ keywords = ["vulkan", "video", "wgpu", "decoding", "encoding"]
 
 [features]
 expose-parsers = []
+dmabuf = ["wgpu"]
+vaapi = ["dmabuf", "dep:crossbeam-channel", "dep:libva"]
 vk-api-dump = []
 vk-validation = []
 wgpu = ["dep:wgpu"]
@@ -30,6 +32,7 @@ default = ["wgpu"]
 [dependencies]
 bytemuck = { version = "1.19.0", features = ["derive"], optional = true }
 bytes = "1"
+crossbeam-channel = { version = "0.5.15", optional = true }
 derivative = "2.2.0"
 h264-reader = { workspace = true }
 memchr = "2.7.4"
@@ -43,12 +46,16 @@ wgpu = { git = "https://github.com/gfx-rs/wgpu", rev = "1503796", optional = tru
 ash = "0.38.0"
 vk-mem = "0.4.0"
 
+[target.'cfg(target_os = "linux")'.dependencies]
+libva = { package = "cros-libva", git = "https://github.com/michael-dm/cros-libva", branch = "export-prime-options", optional = true }
+
 [dev-dependencies]
 clap = { version = "4.5.20", features = ["derive"] }
 tracing-subscriber = "0.3.20"
 wgpu = { git = "https://github.com/gfx-rs/wgpu", rev = "1503796", features = ["noop"] }
 winit = "0.29"
 bytemuck = { version = "1.19.0", features = ["derive"] }
+pollster = { workspace = true }
 
 [build-dependencies]
 cfg_aliases = "0.2.1"
