+++ gpu-video/src/parser.rs
@@ -2,14 +2,10 @@ mod au_splitter;
 mod nalu_parser;
 mod nalu_splitter;
 
-#[cfg(all(vulkan, not(feature = "expose-parsers")))]
+#[cfg(vulkan)]
 pub(crate) mod decoder_instructions;
-#[cfg(feature = "expose-parsers")]
-pub mod decoder_instructions;
-#[cfg(all(vulkan, not(feature = "expose-parsers")))]
+#[cfg(vulkan)]
 pub(crate) mod reference_manager;
-#[cfg(feature = "expose-parsers")]
-pub mod reference_manager;
 
 pub mod h264 {
     use super::au_splitter::AUSplitter;
