+++ gpu-video/src/vulkan_video/wgpu_api.rs
@@ -58,6 +58,7 @@ impl WgpuTexturesDecoder {
             }
             DecoderEvent::SignalDataLoss => {
                 self.reference_ctx.mark_missed_frames();
+                self.frame_sorter.clear();
                 Ok(Vec::new())
             }
             DecoderEvent::Flush => {
