+++ smelter-render/Cargo.toml
@@ -9,7 +9,6 @@ license-file = "../LICENSE"
 web-renderer = ["dep:libcef", "dep:shared_memory"]
 
 [dependencies]
-ash = "0.38.0"
 pollster = { workspace = true }
 thiserror = { workspace = true }
 wgpu = { workspace = true }
@@ -28,6 +27,9 @@ tracing = { workspace = true }
 shared_memory = { workspace = true, optional = true }
 itertools = { workspace = true }
 sys-locale = "0.3.1"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+gpu-video = { workspace = true, features = ["dmabuf"] }
 
 [lints]
 workspace = true
